# openapi.yml
openapi: 3.0.3
info:
  title: ClubManager API
  description: >
    Specification for the ClubManager API,
    used by web and mobile clients to interact with the server.
  version: 0.1.0
...
components:
  securitySchemes:
  # I'm not sure if we are going to need BasicAuth depends how we handle Cognito
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: [] 
paths:
...
  /api/member/new:
    post:
      tags:
        - member
      summary: Create a new member
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/member/{memberID}:
    parameters:
      - name: memberID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - member
      summary: Get this member
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - member
      summary: Update this member
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - member
      summary: Delete this member
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/member/list:
    get:
      tags:
        - member
      summary: Get all members
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: role
          in: query
          description: "Optional: filter members by a specific role"
          schema:
            type: string
            enum:
              - boardMember
              - member
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO

  # membership
  /api/membership/new:
    post:
      tags:
        - membership
      summary: Create a new membership
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/membership/{membershipID}:
    parameters:
      - name: membershipID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - membership
      summary: Get this membership
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - membership
      summary: Update this membership
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.

        **Note:** This endpoint is used to mark a membership as inactive, rather
        than using a `DELETE` endpoint to remove it entirely.
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/membership/list:
    get:
      tags:
        - membership
      summary: Get all memberships
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: status
          in: query
          description: "Optional: filter memberships by a specific status"
          schema:
            type: string
            enum: 
              - active
              - inactive
              - pending
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/membership/register:
    post:
      tags:
        - membership
      summary: Register for a new membership
      security: {}
      description: This request does not require authentication.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
        
  # bike
...